+++ pkg/controller/oneagent/oneagent_controller.go
@@ -438,7 +438,7 @@ func newPodSpecForCR(instance *dynatracev1alpha1.OneAgent) corev1.PodSpec {
 		HostIPC:            true,
 		NodeSelector:       instance.Spec.NodeSelector,
 		PriorityClassName:  instance.Spec.PriorityClassName,
-		ServiceAccountName: "dynatrace-oneagent",
+		ServiceAccountName: instance.Spec.ServiceAccountName,
 		Tolerations:        instance.Spec.Tolerations,
 		DNSPolicy:          instance.Spec.DNSPolicy,
 		Volumes: []corev1.Volume{{