+++ pkg/controller/oneagent/oneagent_utils.go
@@ -83,6 +83,8 @@ func copyDaemonSetSpecToOneAgentSpec(dsSpec *appsv1.DaemonSetSpec, crSpec *dynat
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
+
+	crSpec.ServiceAccountName = dsSpec.Template.Spec.ServiceAccountName
 	crSpec.PriorityClassName = dsSpec.Template.Spec.PriorityClassName
 	crSpec.DNSPolicy = dsSpec.Template.Spec.DNSPolicy
 	crSpec.Image = ""